refactor: separation of function in files
