+++ src/components/ProjectIcons.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { IconsData } from '../hooks/useGithubAutomatedRepos';
+import { projectIconsURL } from '../icons/projectIconsURL';
 
 type Props = {
     itemTopics: string;
@@ -10,15 +10,15 @@ type Props = {
 /**
  * @param {string} itemTopics - Mandatory: ex.: item.topics.map(topics)...
  * @param {string} className - Optional: style className.
- * @returns {<img/>} - Return tag img, project svg icon.
+ * @returns {ReactNode} - Return tag img, project svg icon.
  */
 
 export function ProjectIcons({ itemTopics, className }: Props) {
-    const { projectIcons } = IconsData();
+   
     return (
         <div>
-            {projectIcons[itemTopics as never] ? (
-                <img className={className} alt={projectIcons[itemTopics as never]} src={projectIcons[itemTopics as never]} />
+            {projectIconsURL[itemTopics] ? (
+                <img className={className} alt={projectIconsURL[itemTopics]} src={projectIconsURL[itemTopics]} />
             ) : (
                 <> </>
             )}
