+++ src/icons/projectIconsURL.ts
@@ -1,0 +1,38 @@
+  // Project Icon Font:
+// https://www.svgrepo.com/svg/384245/controller-joystick-games-video-console?edit=true
+// https://www.svgrepo.com/svg/383702/landing-page-web-design?edit=true
+// https://www.svgrepo.com/svg/459107/robot?edit=true
+// https://www.svgrepo.com/svg/471159/check-square-broken?edit=true
+// https://www.svgrepo.com/svg/383754/security-protection-lock-padlock-square-locked?edit=true
+// https://www.svgrepo.com/svg/339111/dashboard?edit=true
+// https://www.svgrepo.com/svg/257454/profile?edit=true
+// https://www.svgrepo.com/svg/298828/store?edit=true
+// https://www.svgrepo.com/svg/429278/cart-essential-shopping?edit=true
+// https://www.svgrepo.com/svg/144539/artist-color-palette?edit=true
+// https://www.svgrepo.com/svg/424981/online?edit=true
+// https://www.svgrepo.com/svg/384295/weather-forecast-sunny-snow-cloud-day?edit=true
+// https://www.svgrepo.com/svg/456864/diagram
+// https://www.svgrepo.com/svg/455166/map-navigation?edit=true
+// https://www.svgrepo.com/svg/343466/news-feed?edit=true
+// https://www.svgrepo.com/svg/458454/database-1?edit=true
+
+interface IProjectIcons {
+  readonly [key: string]: string;
+}
+
+  // 64px
+export const projectIconsURL:IProjectIcons = {
+    art: 'https://user-images.githubusercontent.com/59892368/212994060-8d1644c7-96d7-4f3b-8e94-65ff76db0c92.svg',
+    artificialintelligence: 'https://user-images.githubusercontent.com/59892368/212984565-a424b06e-db5e-464a-a5df-ddf7c9bab6ad.svg',
+    dashboard: 'https://user-images.githubusercontent.com/59892368/212991791-588a6c13-795e-47aa-b496-8bdbaa3cac30.svg',
+    education: 'https://user-images.githubusercontent.com/59892368/212284904-0b4f29fa-4141-4212-b516-d4e4fcdb8519.svg',
+    game: 'https://user-images.githubusercontent.com/59892368/212984732-79dd44aa-23a5-4b83-ae28-7a2ddd443032.svg',
+    landingpage: 'https://user-images.githubusercontent.com/59892368/212984364-88a0808b-a8d1-440c-b74e-f82bc3e28748.svg',
+    personalwebsite: 'https://user-images.githubusercontent.com/59892368/213131421-fa77f07c-9120-4b40-859d-2bd799de6681.svg',
+    productivity: 'https://user-images.githubusercontent.com/59892368/212994727-f05595d9-53b7-43a5-b974-b27fa2557653.svg',
+    security: 'https://user-images.githubusercontent.com/59892368/212985266-a658da1c-64ba-46b7-a069-a148ca217be9.svg',
+    store: 'https://user-images.githubusercontent.com/59892368/213137554-83aa6798-9487-4d1b-a260-fd2035adaaad.svg',
+};
+
+
+
