+++ src/index.ts
@@ -1,4 +1,4 @@
 export { ProjectIcons } from './components/ProjectIcons';
 export { StackIcons } from './components/StackIcons';
 export { StackLabels } from './components/StackLabels/StackLabels';
-export { IGithubRepos, useGitHubAutomatedRepos, IconsData } from './hooks/useGithubAutomatedRepos';
+export { IGithubRepos, useGitHubAutomatedRepos} from './hook/useGithubAutomatedRepos';
